refactor(server): change wallet lock management amounts from double to long for consistency

+++ server/src/main/java/database/dao/WalletDAO.java
@@ -96,7 +96,7 @@ public class WalletDAO {
     /**
      * Chuyển tiền từ số dư khả dụng sang tạm giữ (Locking)
      */
-    public boolean lockBalance(Connection conn, String userId, double amount) throws SQLException {
+    public boolean lockBalance(Connection conn, String userId, long amount) throws SQLException {
         String sql = "UPDATE wallets SET balance = balance - ?, locked_balance = locked_balance + ? " +
                 "WHERE user_id = ? AND balance >= ?";
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -111,7 +111,7 @@ public class WalletDAO {
     /**
      * Hoàn trả tiền từ tạm giữ về số dư khả dụng (Unlocking)
      */
-    public boolean unlockBalance(Connection conn, String userId, double amount) throws SQLException {
+    public boolean unlockBalance(Connection conn, String userId, long amount) throws SQLException {
         String sql = "UPDATE wallets SET balance = balance + ?, locked_balance = locked_balance - ? " +
                 "WHERE user_id = ? AND locked_balance >= ?";
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -126,7 +126,7 @@ public class WalletDAO {
     /**
      * Khấu trừ vĩnh viễn từ số tiền đã tạm giữ (Deducting from lock)
      */
-    public boolean deductFromLocked(Connection conn, String userId, double amount) throws SQLException {
+    public boolean deductFromLocked(Connection conn, String userId, long amount) throws SQLException {
         String sql = "UPDATE wallets SET locked_balance = locked_balance - ? " +
                 "WHERE user_id = ? AND locked_balance >= ?";
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
